Improve parsing of insecure option
According to the git-config manpage, a boolean option may have the
following formats:

When a variable is said to take a boolean value, many synonyms are
accepted for true and false; these are all case-insensitive.

true
   Boolean true can be spelled as yes, on, true, or 1. Also, a variable
   defined without = <value> is taken as true.

false
   Boolean false can be spelled as no, off, false, or 0.

This implements every value of true except if its just the variable
without any value. Everything else will be parsed as false.